ultima entrega 1

+++ src/controllers/product.controller.js
@@ -1,11 +1,7 @@
-import { routes } from "../utils.js";
-// import ProductManager from "../dao/Mongo/productManagerMongo.js";
 import { productsModel } from "../dao/models/products.model.js";
 import { productRepository } from "../services/services.js";
 import { usersModel } from "../dao/models/users.model.js";
-
-const rutaProductos = routes.products;
-// const products = new ProductManager(rutaProductos);
+import { enviarMail } from "../config/mailer.js";
 
 export const getAllProducts = async (req, res) => {
   let { pagina, limit, query, sort } = req.query;
@@ -107,9 +103,32 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const productId = req.params.pid;
+
+    const product = await productRepository.getProductById(productId);
+    if (!product) {
+      return res.status(404).json({ error: "Producto no encontrado" });
+    }
+
+    const owner = await usersModel.findOne({ email: product.owner });
+    if (!owner) {
+      return res.status(404).json({ error: "Propietario del producto no encontrado" });
+    }
+
+    if (owner.rol === "premium") {
+      const subject = "Producto eliminado";
+      const message = `
+              <p>Hola ${owner.first_name},</p>
+              <p>Le informamos que su producto <strong>${product.title}</strong> ha sido eliminado de nuestra plataforma.</p>
+              <p>Saludos,</p>
+              <p>El equipo de E-commerce</p>
+          `;
+      await enviarMail(owner.email, subject, message);
+    }
+
     await productRepository.deleteProduct(productId);
     res.json({ message: "Producto eliminado correctamente" });
   } catch (error) {
-    res.status(404).json({ error: error.message });
+    req.logger.error(error);
+    res.status(500).json({ error: error.message });
   }
 };
